Check protoc exist (#4463)

+++ scripts/build_proto.sh
@@ -3,6 +3,12 @@ CLEAN="src/syft/proto"
 PYTHON_OUT="src/syft"
 PROTO_IN="proto"
 
+command -v protoc &> /dev/null
+if [ $? -ne 0 ]; then
+    echo -ne "Install protobuf\n"
+    exit 1
+fi
+
 rm -rf "${CLEAN}"
 find ${PROTO_IN} -name "*.proto" -print0 | xargs -0 protoc --python_out=${PYTHON_OUT}
